Demo redesign, links to npm and stackblitz
