+++ projects/demo/src/app/app.component.html
@@ -1,185 +1,13 @@
-<header class="mat-elevation-z2">
-  <mat-toolbar>
-    <h2>ngx-imagery</h2>
-    <a href="https://github.com/daelmaak/ngx-imagery" class="github-link">
-      <img src="./assets/GitHub-Mark-64px.png" alt="github link" />
-    </a>
-  </mat-toolbar>
-</header>
+<app-header></app-header>
+
+<mat-menu #demosMenu="matMenu">
+  <a routerLink="/demos/custom-item-template" mat-menu-item
+    >Custom item template</a
+  >
+</mat-menu>
 
 <main>
-  <form>
-    <section>
-      <mat-checkbox name="arrows" [(ngModel)]="arrows">Arrows</mat-checkbox>
-    </section>
-
-    <section>
-      <mat-checkbox name="imageCounter" [(ngModel)]="imageCounter"
-        >Image counter</mat-checkbox
-      >
-    </section>
-
-    <section>
-      <label>Image counter orientation:</label>
-      <mat-radio-group
-        name="imageCounterOrientation"
-        [(ngModel)]="imageCounterOrientation"
-      >
-        <mat-radio-button value="top">top</mat-radio-button>
-        <mat-radio-button value="bottom">bottom</mat-radio-button>
-      </mat-radio-group>
-    </section>
-
-    <section>
-      <label>Image fit:</label>
-      <mat-radio-group name="imageFit" [(ngModel)]="imageFit">
-        <mat-radio-button value="contain">contain</mat-radio-button>
-        <mat-radio-button value="cover">cover</mat-radio-button>
-      </mat-radio-group>
-    </section>
-
-    <section>
-      <mat-checkbox name="loop" [(ngModel)]="loop">Loop</mat-checkbox>
-    </section>
-
-    <section>
-      <mat-checkbox name="thumbs" [(ngModel)]="thumbs">Thumbnails</mat-checkbox>
-    </section>
-
-    <section>
-      <mat-checkbox name="thumbsAutoScroll" [(ngModel)]="thumbsAutoScroll"
-        >Thumbs autoscroll</mat-checkbox
-      >
-    </section>
-
-    <section>
-      <label>Thumbnails orientation:</label>
-      <mat-radio-group name="thumbsOrientation" [(ngModel)]="thumbsOrientation">
-        <mat-radio-button value="top">top</mat-radio-button>
-        <mat-radio-button value="bottom">bottom</mat-radio-button>
-        <mat-radio-button value="left">left</mat-radio-button>
-        <mat-radio-button value="right">right</mat-radio-button>
-      </mat-radio-group>
-    </section>
-
-    <section>
-      <mat-checkbox name="thumbsArrows" [(ngModel)]="thumbsArrows"
-        >Thumbnail arrows</mat-checkbox
-      >
-      <mat-form-field>
-        <label>Slide thumbnails by length</label>
-        <input
-          name="thumbsArrowsSlideByLength"
-          matInput
-          [(ngModel)]="thumbsArrowSlideByLength"
-        />
-      </mat-form-field>
-    </section>
-
-    <section>
-      <label>Thumbnails scroll behavior:</label>
-      <mat-radio-group
-        name="thumbsScrollBehavior"
-        [(ngModel)]="thumbsScrollBehavior"
-      >
-        <mat-radio-button value="smooth">smooth</mat-radio-button>
-        <mat-radio-button value="auto">auto</mat-radio-button>
-      </mat-radio-group>
-    </section>
-
-    <section>
-      <label>Thumbnails overscroll behavior:</label>
-      <mat-radio-group
-        name="thumbsOverscrollBehavior"
-        [(ngModel)]="thumbsOverscrollBehavior"
-      >
-        <mat-radio-button value="contain">contain</mat-radio-button>
-        <mat-radio-button value="auto">auto</mat-radio-button>
-      </mat-radio-group>
-    </section>
-
-    <section>
-      <label>Thumbnail image fit:</label>
-      <mat-radio-group name="thumbsImageFit" [(ngModel)]="thumbsImageFit">
-        <mat-radio-button value="contain">contain</mat-radio-button>
-        <mat-radio-button value="cover">cover</mat-radio-button>
-      </mat-radio-group>
-    </section>
-
-    <section>
-      <label>Image loading strategy:</label>
-      <mat-radio-group name="loading" [(ngModel)]="loading">
-        <mat-radio-button value="auto">auto</mat-radio-button>
-        <mat-radio-button value="lazy">lazy</mat-radio-button>
-      </mat-radio-group>
-      <button mat-raised-button (click)="reloadGallery()">
-        Confirm loading strategy
-      </button>
-    </section>
-
-    <hr />
-
-    <section class="image-loading-latency">
-      <mat-form-field>
-        <label>Items (not images!) loading latency</label>
-        <input
-          name="imageLoadingLatency"
-          matInput
-          [(ngModel)]="imageLoadingLatency"
-        />
-      </mat-form-field>
-      <button mat-raised-button (click)="reloadGallery()">
-        Confirm latency
-      </button>
-    </section>
-  </form>
-
-  <ngx-gallery
-    #gallery
-    class="first"
-    *ngIf="displayGallery"
-    [items]="images | async"
-    [arrows]="arrows"
-    [imageCounter]="imageCounter"
-    [imageCounterOrientation]="imageCounterOrientation"
-    [imageFit]="imageFit"
-    [loading]="loading"
-    [loop]="loop"
-    [thumbs]="thumbs"
-    [thumbsAutoScroll]="thumbsAutoScroll"
-    [thumbsOrientation]="thumbsOrientation"
-    [thumbsArrows]="thumbsArrows"
-    [thumbsArrowSlideByLength]="thumbsArrowSlideByLength"
-    [thumbsScrollBehavior]="thumbsScrollBehavior"
-    [thumbsOverscrollBehavior]="thumbsOverscrollBehavior"
-    [thumbsImageFit]="thumbsImageFit"
-    (imageClick)="openFirst(gallery.selectedIndex)"
-  ></ngx-gallery>
+  <router-outlet></router-outlet>
 </main>
 
-<ng-template
-  #itemTemplate
-  let-index="index"
-  let-selectedIndex="selectedIndex"
-  let-item="item"
-  let-visited="visited"
->
-  <img
-    #img
-    class="custom-image"
-    [src]="visited ? item.src : ''"
-    (load)="onImageLoad()"
-  />
-  <div class="loading" *ngIf="visited && img.naturalHeight <= 0">
-    Loading ...
-  </div>
-</ng-template>
-
-<ng-template
-  #thumbTemplate
-  let-index="index"
-  let-selectedIndex="selectedIndex"
-  let-item="item"
->
-  <img class="custom-thumb" [src]="item.thumbSrc || item.src" />
-</ng-template>
+<footer>With care from daelmaak</footer>
