+++ projects/demo/src/app/app.module.ts
@@ -3,32 +3,49 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
+import { RouterModule, Route } from '@angular/router';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCheckboxModule } from '@angular/material/checkbox';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
+import { MatMenuModule } from '@angular/material/menu';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatRadioModule } from '@angular/material/radio';
 
 import { AppComponent } from './app.component';
 import { GalleryModule } from 'projects/gallery/src/public-api';
 import { GalleryDetailModule } from 'projects/gallery-detail/src/public-api';
+import { HomeComponent } from './pages/home/home.component';
+import { HeaderComponent } from './components/header/header.component';
+
+const materialModules = [
+  MatButtonModule,
+  MatCheckboxModule,
+  MatFormFieldModule,
+  MatInputModule,
+  MatMenuModule,
+  MatRadioModule,
+  MatToolbarModule
+];
+
+const routes: Route[] = [
+  {
+    path: '',
+    component: HomeComponent
+  }
+];
 
 @NgModule({
-  declarations: [AppComponent],
+  declarations: [AppComponent, HomeComponent, HeaderComponent],
   imports: [
+    ...materialModules,
     BrowserModule,
     BrowserAnimationsModule,
     CommonModule,
     FormsModule,
     GalleryModule,
     GalleryDetailModule,
-    MatButtonModule,
-    MatCheckboxModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatRadioModule,
-    MatToolbarModule
+    RouterModule.forRoot(routes)
   ],
   providers: [],
   bootstrap: [AppComponent]
