+++ projects/demo/src/app/components/header/header.component.html
@@ -1,0 +1,30 @@
+<header class="mat-elevation-z2">
+  <mat-toolbar>
+    <a routerLink="/" class="app-title"><h2>ngx-imagery</h2></a>
+
+    <div class="links">
+      <a
+        href="https://www.npmjs.com/package/@ngx-imagery/gallery"
+        class="npm-link link"
+      >
+        <img src="./assets/icons/npm-logo.svg" alt="npm link" />
+      </a>
+      <a
+        href="https://stackblitz.com/edit/ngx-imagery"
+        class="stackblitz-link link"
+        target="_blank"
+      >
+        <img src="./assets/icons/stackblitz-logo.svg" alt="stackblitz link" />
+      </a>
+      <a
+        href="https://github.com/daelmaak/ngx-imagery"
+        class="github-link link"
+      >
+        <img
+          src="./assets/icons/GitHub-Mark-Light-64px.png"
+          alt="github link"
+        />
+      </a>
+    </div>
+  </mat-toolbar>
+</header>
